test: Convert sleep millis to `u32` to avoid overflow of `DateTime`

+++ crates/testing/test-programs/sleep/workflow/src/lib.rs
@@ -7,11 +7,11 @@ bindings::export!(Component with_types_in bindings);
 struct Component;
 
 impl crate::bindings::exports::testing::sleep_workflow::workflow::Guest for Component {
-    fn sleep_host_activity(millis: u64) {
+    fn sleep_host_activity(millis: u32) {
         crate::bindings::my_org::workflow_engine::host_activities::sleep(millis);
     }
 
-    fn sleep_activity(millis: u64) {
+    fn sleep_activity(millis: u32) {
         crate::bindings::testing::sleep::sleep::sleep(millis);
     }
